backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Firebase and Google Cloud
firebase-admin==6.2.0
google-cloud-storage>=2.10.0
google-generativeai>=0.3.0

# HTTP clients
requests==2.31.0
httpx>=0.25.0

# Async job processing and task queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Rate limiting and throttling
slowapi==0.1.9
limits==3.6.0

# Database connection pooling
psycopg2-binary==2.9.9  # For PostgreSQL if needed
sqlalchemy==2.0.23
asyncpg==0.29.0  # Async PostgreSQL driver

# Caching
aioredis>=2.0.1
cachetools==5.3.2

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Background task management
aiofiles==23.2.1

# Circuit breaker pattern
circuitbreaker==1.4.0

# Retry mechanisms
tenacity==8.2.3

# Load balancing and health checks
healthcheck==1.3.3
